♻️ Separated helping functions from route file
